base64c: change base64_valid to base64_verify
This is not just name change but also change how function works. It now
returns zero when input is base64 encoded string. Error index is
provided as well but because of collision with zero it is shifted by
one. That means that base64_verify returns 1 if first character is
invalid.

The reason for this change is that it simplifies usage of function if
index of invalid character is not of the interest but rather plain
verification is required.